chore(client):extract upgrade button to a separate component
